use inline file test

+++ scripts/openssl.sh
@@ -47,8 +47,5 @@ rm -rf /usr/include/crypto
 
 cp -r -L ./include/openssl /usr/include
 # include/crypto only exists in some OpenSSL versions
-if [ -e ./include/crypto ]
-then
-    cp -r -L ./include/crypto /usr/include
-fi
+test -e ./include/crypto && cp -r -L ./include/crypto /usr/include
 cp -H ./libcrypto.so "/usr/lib/libcrypto.so.${version}"
